refactor: remove unused exports
Also moved a test utilities file to the folder in which it is used
