+++ viewer/packages/cad-geometry-loaders/index.ts
@@ -4,19 +4,12 @@
 
 export {
   CadModelBudget,
-  defaultCadModelBudget,
   defaultDesktopCadModelBudget,
-  defaultMobileCadModelBudget
 } from './src/CadModelBudget';
-
-export { CadLoadingHints } from './src/CadLoadingHints';
 
 export { CadModelUpdateHandler } from './src/CadModelUpdateHandler';
 
 export { SectorCuller } from './src/sector/culling/SectorCuller';
 export { ByVisibilityGpuSectorCuller } from './src/sector/culling/ByVisibilityGpuSectorCuller';
-export { ByScreenSizeSectorCuller } from './src/sector/culling/ByScreenSizeSectorCuller';
 
 export { createV8SectorCuller } from './src/sector/culling/createV8SectorCuller';
-
-export { DetermineSectorsInput } from './src/sector/culling/types';
