+++ viewer/packages/cad-parsers/index.ts
@@ -2,10 +2,8 @@
  * Copyright 2021 Cognite AS
  */
 
-export { CadMetadataParser } from './src/metadata/CadMetadataParser';
 export { CadModelMetadataRepository } from './src/metadata/CadModelMetadataRepository';
 export { CadModelMetadata } from './src/metadata/CadModelMetadata';
-export { SectorSceneImpl } from './src/utilities/SectorScene';
 export { SectorSceneFactory } from './src/utilities/SectorSceneFactory';
 export { CadModelClipper } from './src/metadata/CadModelClipper';
