+++ viewer/packages/metrics/index.ts
@@ -2,5 +2,3 @@
  * Copyright 2021 Cognite AS
  */
 export { MetricsLogger } from './src/MetricsLogger';
-
-export { EventProps, TrackedEvents } from './src/types';
