+++ viewer/packages/rendering/index.ts
@@ -8,11 +8,7 @@ export { CadMaterialManager } from './src/CadMaterialManager';
 
 export { createSimpleGeometryMesh } from './src/rendering/createSimpleGeometryMesh';
 
-export { NodeAppearanceTextureBuilder } from './src/rendering/NodeAppearanceTextureBuilder';
-export { NodeTransformTextureBuilder } from './src/transform/NodeTransformTextureBuilder';
 export { NodeTransformProvider } from './src/transform/NodeTransformProvider';
-
-export { CogniteColors, RevealColors } from './src/utilities/types';
 
 export { CadNode } from './src/sector/CadNode';
 export {
@@ -28,5 +24,5 @@ export { EffectRenderManager } from './src/rendering/EffectRenderManager';
 export { RenderMode } from './src/rendering/RenderMode';
 export { coverageShaders } from './src/rendering/shaders';
 
-export { Materials, createMaterials } from './src/rendering/materials';
+export { Materials } from './src/rendering/materials';
 export { createPrimitives } from './src/rendering/primitives';
