+++ viewer/test-utilities/index.ts
@@ -14,7 +14,6 @@ export { createV8SceneSectorMetadata, createV9SceneSectorMetadata } from './src/
 
 export { createCadModelMetadata } from './src/createCadModelMetadata';
 export { createCadModel } from './src/createCadModel';
-export { createDetermineSectorInput } from './src/createDetermineSectorInput';
 
 export { mockClientAuthentication } from './src/cogniteClientAuth';
